+++ devapp_v1/routes/webpage_routes.js
@@ -1,0 +1,24 @@
+var express = require('express');
+var router = express.Router();
+
+// Homepage
+router.get('/', function(req, res) {
+
+});
+
+// Login page
+router.get('/login', function(req,res) {
+
+});
+
+// About page
+router.get('/about', function(req, res) {
+
+});
+
+// Edit page
+router.get('/edit', function(req, res) {
+
+});
+
+module.exports = router;
